requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.2
numpy==1.26.4
openai==1.25.1
pandas==2.2.2
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
